Fix int to double

+++ src/main/java/it/unibo/pensilina14/bullet/ballet/input/Right.java
@@ -9,7 +9,7 @@ public class Right implements Command {
 	 * Factor every time Right command is executed
 	 * the player moves by.
 	 */
-	public static final int MOVEMENT_DELTA = 0.1;
+	public static final double MOVEMENT_DELTA = 0.1;
 	
 	@Override
 	public final void execute(final GameState env) {
